% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_ball.R
\name{get_ctheta_samps}
\alias{get_ctheta_samps}
\title{Posterior samples of \code{c_theta}.}
\usage{
get_ctheta_samps(theta, omega, p, max.k)
}
\arguments{
\item{theta}{A list of MCMC samples from the posterior of the localized atoms.}

\item{omega}{The FOLD separation parameter, positive.}

\item{p}{Number of features.}

\item{max.k}{Maximum number of clusters.}
}
\value{
A matrix with rows corresponding to samples from the posterior of \code{c_theta}.
}
\description{
Posterior samples of \code{c_theta}.
}
\details{
This function extracts posterior samples from \code{c_theta}, a minimizer of the FOLD
loss function. The resulting matrix of samples can then be used to compute the credible ball.
}
