% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_gibbs.R
\name{lmvnorm_gibbs}
\alias{lmvnorm_gibbs}
\title{Gibbs sampler for a location Bayesian GMM with fixed covariance}
\usage{
lmvnorm_gibbs(S, y, L, alpha = rep(1/L, L), mu_0, Sigma_0, Sigma, stops = 1)
}
\arguments{
\item{S}{Total number of MCMC iterations.}

\item{y}{Matrix with \code{n} observations as the rows and \code{p} features as the columns.}

\item{L}{Number of mixture components. Default is \code{L=10}.}

\item{alpha}{Vector of Dirichlet concentration parameters. Default is \code{(1/L, ..., 1/L)}.}

\item{mu_0}{Prior mean for component centers.}

\item{Sigma_0}{Prior covariance for component centers.}

\item{Sigma}{Covariance of the data, fixed and constant across components.}

\item{stops}{Number of iterations to print progress point.}
}
\value{
A list MCMC samples from the posteriors of: \code{theta} (localized centers), and \code{z} (component labels).
}
\description{
Samples from the posterior distribution of a location GMM where the covariance is fixed and constant.
}
\details{
This function fits a special case of the Bayesian GMM in which the component-specific covariance \code{Sigma} is
fixed and constant across all components. The cluster means, or centers, are drawn from a normal distribution with parameters
\code{mu_0} and \code{Sigma_0}, while the weights are given a Dirichlet prior with parameter \code{alpha}.
}
