% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{min_risk}
\alias{min_risk}
\title{Minimize FOLD risk function}
\usage{
min_risk(c, Delta, omega)
}
\arguments{
\item{c}{A matrix, each row is a vector of cluster labels.}

\item{Delta}{An \code{n}x\code{n} symmetric matrix of pairwise expected Hellinger distances.}

\item{omega}{FOLD separation parameter, positive.}
}
\value{
A vector, the clustering which minimizes the FOLD risk.
}
\description{
Minimizes the FOLD risk function over a matrix of clusterings.
}
