% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_gibbs.R
\name{mvnorm_gibbs}
\alias{mvnorm_gibbs}
\title{Gibbs sampler for a multivariate Bayesian GMM}
\usage{
mvnorm_gibbs(S, y, L = 10, alpha = rep(1/L, L), w, kappa, r, C, stops = 1)
}
\arguments{
\item{S}{Total number of MCMC iterations.}

\item{y}{Matrix with \code{n} observations as the rows and \code{p} features as the columns.}

\item{L}{Number of mixture components. Default is \code{L=10}.}

\item{alpha}{Vector of Dirichlet concentration parameters. Default is \code{(1/L, ..., 1/L)}.}

\item{w}{Normal-Inverse-Wishart location parameter, \code{p}-dimensional.}

\item{kappa}{Normal-Inverse-Wishart scale parameter.}

\item{r}{Inverse-Wishart degrees of freedom, greater than \code{p}-1.}

\item{C}{Inverse-Wishart scale matrix, \code{p}x\code{p}.}

\item{stops}{Number of iterations to print progress point.}
}
\value{
A list MCMC samples from the posteriors of: \code{theta} (localized atoms), and \code{z} (component labels).
}
\description{
Samples from the posterior distribution of a GMM with unknown means and covariances.
}
\details{
This function fits a Bayesian GMM in order to cluster a sample of observations. The prior distributions are
conjugate; the means and covariance are drawn from a normal-Inverse-Wishart distribution with parameters \code{w},
\code{kappa}, \code{r}, and \code{C}, while the weights follow a Dirichlet distribution with parameter \code{alpha}.
}
