% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_gibbs.R
\name{unorm_gibbs}
\alias{unorm_gibbs}
\title{Gibbs sampler for a univariate Bayesian GMM}
\usage{
unorm_gibbs(S, y, L = 10, alpha = rep(1/L, L), w, kappa, r, C, stops = 1)
}
\arguments{
\item{S}{Total number of MCMC iterations.}

\item{y}{Vector with \code{n} observations to cluster.}

\item{L}{Number of mixture components. Default is \code{L=10}.}

\item{alpha}{Vector of Dirichlet concentration parameters. Default is \code{(1/L, ..., 1/L)}.}

\item{w}{Normal-Inverse-Gamma location parameter, scalar.}

\item{kappa}{Normal-Inverse-Gamma scale parameter, positive.}

\item{r}{Inverse-Gamma degrees of freedom, positive.}

\item{C}{Inverse-Gamma scale, positive.}

\item{stops}{Number of iterations to print progress point.}
}
\value{
A list MCMC samples from the posteriors of: \code{theta} (localized atoms), and \code{z} (component labels).
}
\description{
Samples from the posterior distribution of a GMM with unknown means and variances.
}
\details{
This function fits a univariate Bayesian GMM in order to cluståer a sample of observations. The prior distributions are
conjugate; the means and variances are drawn from a normal-Inverse-Gamma distribution with parameters \code{w},
\code{kappa}, \code{r}, and \code{C}. Note that we have parameterized the variance prior as
Inv-Gamma(\code{r}/2,\code{C}/2), in order to be consistent with the normal-Inverse-Wishart for higher dimensions.
The weights follow a Dirichlet distribution with parameter \code{alpha}.
}
